refactor(elm): rename types.ts to schema.ts
Follow project convention of using schema.ts for zod schema files.
